Remove pwr and replace pw with pwr logic
