+++ README.md
@@ -42,9 +42,7 @@ Commands:
   pwned breaches                get all breaches in the system
   pwned dc                      get all data classes in the system
   pwned pa <email>              get all pastes for an account (email address)
-  pwned pw <password>           check a password (plain text or SHA1 hash) for public exposure
-  pwned pwr <password>          securely check a password for public exposure (password will not
-                                be transmitted over the network)
+  pwned pw <password>           securely check a password for public exposure
   pwned search <account|email>  search breaches and pastes for an account (username or email
                                 address)
 
@@ -165,18 +163,11 @@ $ pwned pa nobody@nowhere.com
   EmailCount: 10547
 ```
 
-Check a password to see if it has been exposed in a data breach:
-
-```
-$ pwned pw Password1234
-Oh no — pwned!
-```
-
 [Securely][search-by-range] check a password to see if it has been exposed in a
 data breach:
 
 ```
-$ pwned pwr Password1234
+$ pwned pw Password1234
 Oh no — pwned!
 ```
