+++ src/commands/pw.js
@@ -1,10 +1,10 @@
-import { pwnedPassword } from 'hibp';
+import hasha from 'hasha';
+import { pwnedPasswordRange } from 'hibp';
 import logger from '../utils/logger';
 import spinner from '../utils/spinner';
 
 export const command = 'pw <password>';
-export const describe =
-  'check a password (plain text or SHA1 hash) for public exposure';
+export const describe = 'securely check a password for public exposure';
 
 export const builder /* istanbul ignore next */ = yargs =>
   yargs
@@ -17,47 +17,43 @@ export const builder /* istanbul ignore next */ = yargs =>
       }
       return true;
     })
-    .option('s', {
-      alias: 'sha1',
-      describe: 'password itself is a SHA1 hash (so hash it again)',
-      type: 'boolean',
-      default: false,
-    })
     .option('r', {
       alias: 'raw',
       describe: 'disable the console spinner',
       type: 'boolean',
       default: false,
     })
-    .group(['s', 'r'], 'Command Options:')
+    .group(['r'], 'Command Options:')
     .group(['h', 'v'], 'Global Options:');
 
 /**
  * Fetches the pwned status for the given password, indicating whether or not it
- * has been previously exposed in a breach. Passwords can be plain text or a
- * SHA1 hash. The remote API will automatically attempt to discern between the
- * two, but in the case where the password you wish to check is actually a
- * plain text string containing a hash and you don't want the API to treat it as
- * a hash, you can override the auto detection behavior by setting the isAHash
- * option to true.
+ * has been previously exposed in a breach.
+ *
+ * When a password hash with the same first 5 characters is found in the Pwned
+ * Passwords repository, the API will respond with the suffix of every hash
+ * beginning with the specified prefix, followed by a count of how many times it
+ * appears in the data set. The API consumer can then search the results of the
+ * response for the presence of their source hash and if not found, the password
+ * does not exist in the data set.
  *
  * @param {Object} argv the parsed argv object
- * @param {string} argv.password a password (plain text string or SHA1 hash)
- * @param {boolean} [argv.sha1] the pre-hashed password is a hash (default:
- * false)
+ * @param {string} argv.password a password (plain text)
  * @param {boolean} [argv.raw] disable the console spinner (default: false)
  * @returns {Promise} the resulting Promise where output is rendered
  */
-export const handler = async ({ password, sha1: isAHash, raw }) => {
+export const handler = async ({ password, raw }) => {
   if (!raw && process.stdout.isTTY) {
     spinner.start();
   }
 
   try {
-    const isPwned = await pwnedPassword(password, { isAHash });
+    const hash = hasha(password, { algorithm: 'sha1' }).toUpperCase();
+    const suffixes = await pwnedPasswordRange(hash.slice(0, 5));
     if (!raw && process.stdout.isTTY) {
       spinner.stop(true);
     }
+    const isPwned = suffixes.includes(hash.slice(5));
     logger.log(isPwned ? 'Oh no — pwned!' : 'Good news — no pwnage found!');
   } catch (err) {
     if (!raw && process.stdout.isTTY) {
