+++ src/commands/pw.test.js
@@ -1,5 +1,5 @@
 import * as hibp from 'hibp';
-import { FOUND, NOT_FOUND, ERROR, ERROR_MSG } from '../../testData';
+import { FOUND, NOT_FOUND, SUFFIXES, ERROR, ERROR_MSG } from '../../testData';
 import logger from '../utils/logger';
 import spinner from '../utils/spinner';
 import { handler as pw } from './pw';
@@ -9,12 +9,10 @@ jest.mock('../utils/spinner');
 
 describe('command: pw', () => {
   beforeAll(() => {
-    hibp.pwnedPassword.mockImplementation(async password => {
-      if (password === FOUND) {
-        return true;
-      } else if (password === NOT_FOUND) {
-        return false;
-      } else if (password === ERROR) {
+    hibp.pwnedPasswordRange.mockImplementation(async prefix => {
+      if (prefix !== ERROR) {
+        return SUFFIXES;
+      } else if (prefix === ERROR) {
         throw new Error(ERROR_MSG);
       }
       throw new Error('Unexpected input!');
@@ -22,74 +20,74 @@ describe('command: pw', () => {
   });
 
   it('should call spinner.start (!raw)', () => {
-    pw({ password: NOT_FOUND, sha1: false, raw: false });
+    pw({ password: FOUND, raw: false });
     expect(spinner.start).toHaveBeenCalledTimes(1);
   });
 
   it('should not call spinner.start (raw)', () => {
-    pw({ password: NOT_FOUND, sha1: false, raw: true });
+    pw({ password: FOUND, raw: true });
     expect(spinner.start).toHaveBeenCalledTimes(0);
   });
 
   it('should call spinner.stop (non-error results, !raw)', () => {
     expect(spinner.stop).toHaveBeenCalledTimes(0);
-    return pw({ password: NOT_FOUND, sha1: false, raw: false }).then(() => {
+    return pw({ password: FOUND, raw: false }).then(() => {
       expect(spinner.stop).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should not call spinner.stop (non-error results, raw)', () => {
     expect(spinner.stop).toHaveBeenCalledTimes(0);
-    return pw({ password: NOT_FOUND, sha1: false, raw: true }).then(() => {
+    return pw({ password: FOUND, raw: true }).then(() => {
       expect(spinner.stop).toHaveBeenCalledTimes(0);
     });
   });
 
   it('should call logger.log (found && !raw)', () => {
     expect(logger.log).toHaveBeenCalledTimes(0);
-    return pw({ password: FOUND, sha1: false, raw: false }).then(() => {
+    return pw({ password: FOUND, raw: false }).then(() => {
       expect(logger.log).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should call logger.log (found && raw)', () => {
     expect(logger.log).toHaveBeenCalledTimes(0);
-    return pw({ password: FOUND, sha1: false, raw: true }).then(() => {
+    return pw({ password: FOUND, raw: true }).then(() => {
       expect(logger.log).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should call logger.log (notFound && !raw)', () => {
     expect(logger.log).toHaveBeenCalledTimes(0);
-    return pw({ password: NOT_FOUND, sha1: false, raw: false }).then(() => {
+    return pw({ password: NOT_FOUND, raw: false }).then(() => {
       expect(logger.log).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should call logger.log (notFound && raw)', () => {
     expect(logger.log).toHaveBeenCalledTimes(0);
-    return pw({ password: NOT_FOUND, sha1: false, raw: true }).then(() => {
+    return pw({ password: NOT_FOUND, raw: true }).then(() => {
       expect(logger.log).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should call spinner.stop (error && !raw)', () => {
     expect(spinner.stop).toHaveBeenCalledTimes(0);
-    return pw({ password: ERROR, sha1: false, raw: false }).then(() => {
+    return pw({ prefix: ERROR, raw: false }).then(() => {
       expect(spinner.stop).toHaveBeenCalledTimes(1);
     });
   });
 
   it('should not call spinner.stop (error && raw)', () => {
     expect(spinner.stop).toHaveBeenCalledTimes(0);
-    return pw({ password: ERROR, sha1: false, raw: true }).then(() => {
+    return pw({ prefix: ERROR, raw: true }).then(() => {
       expect(spinner.stop).toHaveBeenCalledTimes(0);
     });
   });
 
   it('should call logger.error (error)', () => {
     expect(logger.error).toHaveBeenCalledTimes(0);
-    return pw({ password: ERROR, sha1: false, raw: false }).then(() => {
+    return pw({ prefix: ERROR, raw: false }).then(() => {
       expect(logger.log).toHaveBeenCalledTimes(0);
       expect(logger.error).toHaveBeenCalledTimes(1);
     });
